adding genre database table

+++ src/main/java/com/smith/gamingLab/table/Game.java
@@ -17,23 +17,9 @@ public class Game {
     private String description;
     private boolean isDigital;
 
-//    //TODO how to get
-//
-//    /*TODO make genre a list of string, list of enum or just a string?
-//     *  Depends on query?
-//     * */
-//    @Column
-//    @Convert(converter = GenreTypeListToStringConverter.class)
-//    private List<GenreType> genres;
-////
-//    //TODO make this into a HashSet instead?
-//    @Column
-//    @Convert(converter = PlayableModeListToStringConverter.class)
-//    private List<PlayableMode> modes;
-
     public Game() {}
 
-    public long getId() { return id; }
+    public int getId() { return id; }
 
     public String getTitle() { return title; }
 
@@ -46,13 +32,6 @@ public class Game {
     public String getDescription() { return description; }
 
     public boolean getIsDigital() { return isDigital; }
-
-//    public List<GenreType> getGenres() {return genres;}
-
-//    public List<PlayableMode> getPlayableModes() { return modes; }
-
-    //setters
-
 
     public void setRating(Rating rating) { this.rating = rating; }
 
@@ -72,30 +51,6 @@ public class Game {
     public void setDescription(String description) { this.description = description; }
 
     public void setIsDigital(boolean isDigital) { this.isDigital = isDigital; }
-
-//    public void setGenres(List<GenreType> genres) {this.genres = genres;}
-
-//    public void addGenre(GenreType genre) { genres.add(genre); }
-
-//    //TODO unit test this
-//    public void deleteGenre(GenreType genre) {
-//        if (genres.contains(genre)) {
-//            genres.remove(genre);
-//        }
-//    }
-//
-//    public void addPlayableMode(PlayableMode mode) { modes.add(mode);}
-//
-//    public void deletePlayableMode(PlayableMode mode) {
-//        if (modes.contains(mode)) { modes.remove(mode); }
-//    }
-
-
-
-
-
-
-
